+++ src/aiida_eon/__init__.py
@@ -1,3 +1,3 @@
 """AiiDA plugin for eOn."""
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"